KernelSU: integrate scope-minimized manual hooks v1.5 - hybridized -k4.9
This refactors original KSU hooks to replace deep kernel function hooks with targeted hooks.
This backports KernelSU pr#1657 and having pr#2084 elements (32-bit sucompat).
It reduces the scope of kernel function interception and still maintains full fucntionality.

This commit is a squash of the following:
*	fs/exec: do_execve: ksu_handle_execve_sucompat hook
*	fs/exec: compat_do_execve: ksu_handle_execve_sucompat hook
	fs/open: sys_faccessat: ksu_handle_faccessat hook
*	fs/stat: sys_newfstatat: ksu_handle_stat hook
*	fs/stat: sys_fstatat64: ksu_handle_stat hook

The rest of the hooks are handled on kprobe
ref:
	kernel: kp_ksud: restore kprobes for early-boot and used-once hooks
	https://github.com/backslashxx/KernelSU/commit/ab4b4b29ae64931dcb1abcf07b1c1685d90e9079

Signed-off-by: backslashxx <118538522+backslashxx@users.noreply.github.com>